24 bit i2c

+++ src/sapphireos/i2s.h
@@ -28,6 +28,6 @@
 void i2s_v_init( void );
 void i2s_v_start( uint16_t sample_rate, uint8_t sample_bits, bool stereo );
 uint32_t i2s_u32_get_count( void );
-uint32_t i2s_u32_get_samples( int16_t *samples, uint16_t max );
+uint32_t i2s_u32_get_samples( int32_t *samples, uint16_t max );
 
 #endif
